feat: add Tab getLocalStorage/setLocalStorage/clearLocalStorage
Ports Zendriver's get_local_storage/set_local_storage, plus a clear
counterpart, and reads the origin from the target URL with Ktor's
`Url.protocolWithAuthority` rather than by slicing the string.

Doing so required fixing the CDP generator first. `DOMStorage.Item` is
declared in the protocol as an array of strings, but every top-level
array type alias was generated as `List<Double>`: the "array" branch of
`CanBeTypeAlias.jsTypeToKType` ignored `items` entirely, unlike the
`TypeOrReference` overload right below it, which resolves them properly.

`getDOMStorageItems` was therefore unusable as generated, since decoding
["key", "value"] into doubles fails at runtime. Two other aliases were
wrong the same way, `Target.TargetFilter` and `DOMSnapshot.ArrayOfStrings`;
the remaining three are arrays of numbers and were right by accident.

Both overloads now share a single `itemsToKType`, so a type alias and a
property describing the same array can no longer drift apart. A type is
also resolved against the domain it was found in rather than the referring
one, so a `$ref` inside its `items` points at the right domain.

Only the three resulting type changes are committed rather than a full
regeneration, which on this machine produces unrelated formatting churn
across 55 files.

+++ cdp/src/commonMain/kotlin/dev/kdriver/cdp/domain/DOMSnapshot.kt
@@ -452,7 +452,7 @@ public class DOMSnapshot(
         /**
          * Attributes of an `Element` node. Flatten name, value pairs.
          */
-        public val attributes: List<List<Double>>? = null,
+        public val attributes: List<List<Int>>? = null,
         /**
          * Only set for textarea elements, contains the text value.
          */
@@ -510,7 +510,7 @@ public class DOMSnapshot(
         /**
          * Array of indexes specifying computed style strings, filtered according to the `computedStyles` parameter passed to `captureSnapshot`.
          */
-        public val styles: List<List<Double>>,
+        public val styles: List<List<Int>>,
         /**
          * The absolute position bounding box.
          */
